+++ packages/web/jest.config.js
@@ -5,12 +5,13 @@ module.exports = {
   moduleNameMapper: {
     '\\.svg$': '<rootDir>/src/__mocks__/svgrMock.ts',
     '^src/(.+)': '<rootDir>/src/$1',
-    '\\.(css|less)$': 'identity-obj-proxy'
+    '\\.(css|less)$': 'identity-obj-proxy',
+    'nft.storage': '<rootDir>/src/__mocks__/nft.storage.ts'
   },
   globals: {
     'ts-jest': {
-      tsconfig: 'tsconfig.test.json',
-    },
+      tsconfig: 'tsconfig.test.json'
+    }
   },
   transformIgnorePatterns: ['/node_modules/(?!react-markdown)/'],
   setupFiles: ['jest-canvas-mock'],
