+++ packages/web/src/__mocks__/nft.storage.ts
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+export const NFTStorage = class {
+  public async storeBlob(file?: File) {
+    return 'cid'
+  }
+}
